introduce animation_defs.h to solve circular dependency between art.h and animation.h
