+++ src/actions.cc
@@ -5,6 +5,7 @@
 
 #include <assert.h>
 
+#include "animation_defs.h"
 #include "animation.h"
 #include "art.h"
 #include "color.h"
