+++ src/animation.cc
@@ -6,6 +6,7 @@
 #include <stdio.h>
 #include <string.h>
 
+#include "animation_defs.h"
 #include "art.h"
 #include "color.h"
 #include "combat.h"
