+++ src/animation_defs.h
@@ -1,0 +1,118 @@
+#ifndef ANIMATION_DEFS_H
+#define ANIMATION_DEFS_H
+
+namespace fallout {
+
+enum AnimationRequestOptions : int {
+    ANIMATION_REQUEST_NONE = 0x00,
+    ANIMATION_REQUEST_UNRESERVED = 0x01,
+    ANIMATION_REQUEST_RESERVED = 0x02,
+    ANIMATION_REQUEST_NO_STAND = 0x04,
+    ANIMATION_REQUEST_PING = 0x100,
+    ANIMATION_REQUEST_INSIGNIFICANT = 0x200,
+};
+
+constexpr inline AnimationRequestOptions operator|(AnimationRequestOptions lhs, AnimationRequestOptions rhs)
+{
+    return static_cast<AnimationRequestOptions>(static_cast<int>(lhs) | static_cast<int>(rhs));
+}
+
+inline AnimationRequestOptions& operator|=(AnimationRequestOptions& lhs, AnimationRequestOptions rhs)
+{
+    return lhs = lhs | rhs;
+}
+
+// Basic animations: 0-19
+// Knockdown and death: 20-35
+// Change positions: 36-37
+// Weapon: 38-47
+// Single-frame death animations (the last frame of knockdown and death animations): 48-63
+enum AnimationType : int {
+    ANIM_INVALID = -1,
+    ANIM_STAND = 0,
+    ANIM_WALK = 1,
+    ANIM_JUMP_BEGIN = 2,
+    ANIM_JUMP_END = 3,
+    ANIM_CLIMB_LADDER = 4,
+    ANIM_FALLING = 5,
+    ANIM_UP_STAIRS_RIGHT = 6,
+    ANIM_UP_STAIRS_LEFT = 7,
+    ANIM_DOWN_STAIRS_RIGHT = 8,
+    ANIM_DOWN_STAIRS_LEFT = 9,
+    ANIM_MAGIC_HANDS_GROUND = 10,
+    ANIM_MAGIC_HANDS_MIDDLE = 11,
+    ANIM_MAGIC_HANDS_UP = 12,
+    ANIM_DODGE_ANIM = 13,
+    ANIM_HIT_FROM_FRONT = 14,
+    ANIM_HIT_FROM_BACK = 15,
+    ANIM_THROW_PUNCH = 16,
+    ANIM_KICK_LEG = 17,
+    ANIM_THROW_ANIM = 18,
+    ANIM_RUNNING = 19,
+    ANIM_FALL_BACK = 20,
+    ANIM_FALL_FRONT = 21,
+    ANIM_BAD_LANDING = 22,
+    ANIM_BIG_HOLE = 23,
+    ANIM_CHARRED_BODY = 24,
+    ANIM_CHUNKS_OF_FLESH = 25,
+    ANIM_DANCING_AUTOFIRE = 26,
+    ANIM_ELECTRIFY = 27,
+    ANIM_SLICED_IN_HALF = 28,
+    ANIM_BURNED_TO_NOTHING = 29,
+    ANIM_ELECTRIFIED_TO_NOTHING = 30,
+    ANIM_EXPLODED_TO_NOTHING = 31,
+    ANIM_MELTED_TO_NOTHING = 32,
+    ANIM_FIRE_DANCE = 33,
+    ANIM_FALL_BACK_BLOOD = 34,
+    ANIM_FALL_FRONT_BLOOD = 35,
+    ANIM_PRONE_TO_STANDING = 36,
+    ANIM_BACK_TO_STANDING = 37,
+    ANIM_TAKE_OUT = 38,
+    ANIM_PUT_AWAY = 39,
+    ANIM_PARRY_ANIM = 40,
+    ANIM_THRUST_ANIM = 41,
+    ANIM_SWING_ANIM = 42,
+    ANIM_POINT = 43,
+    ANIM_UNPOINT = 44,
+    ANIM_FIRE_SINGLE = 45,
+    ANIM_FIRE_BURST = 46,
+    ANIM_FIRE_CONTINUOUS = 47,
+    ANIM_FALL_BACK_SF = 48,
+    ANIM_FALL_FRONT_SF = 49,
+    ANIM_BAD_LANDING_SF = 50,
+    ANIM_BIG_HOLE_SF = 51,
+    ANIM_CHARRED_BODY_SF = 52,
+    ANIM_CHUNKS_OF_FLESH_SF = 53,
+    ANIM_DANCING_AUTOFIRE_SF = 54,
+    ANIM_ELECTRIFY_SF = 55,
+    ANIM_SLICED_IN_HALF_SF = 56,
+    ANIM_BURNED_TO_NOTHING_SF = 57,
+    ANIM_ELECTRIFIED_TO_NOTHING_SF = 58,
+    ANIM_EXPLODED_TO_NOTHING_SF = 59,
+    ANIM_MELTED_TO_NOTHING_SF = 60,
+    ANIM_FIRE_DANCE_SF = 61,
+    ANIM_FALL_BACK_BLOOD_SF = 62,
+    ANIM_FALL_FRONT_BLOOD_SF = 63,
+    ANIM_CALLED_SHOT_PIC = 64,
+    ANIM_COUNT = 65,
+    ANIM_FIRST = ANIM_STAND,
+    FIRST_KNOCKDOWN_AND_DEATH_ANIM = ANIM_FALL_BACK,
+    LAST_KNOCKDOWN_AND_DEATH_ANIM = ANIM_FALL_FRONT_BLOOD,
+    FIRST_SF_DEATH_ANIM = ANIM_FALL_BACK_SF,
+    LAST_SF_DEATH_ANIM = ANIM_FALL_FRONT_BLOOD_SF,
+};
+
+inline bool animationTypeIsValid(int anim)
+{
+    return anim >= ANIM_FIRST && anim < ANIM_COUNT;
+}
+
+inline AnimationType animationTypeFromFid(int fid)
+{
+    int anim = (fid & 0xFF0000) >> 16;
+    return static_cast<AnimationType>(anim);
+}
+
+} // namespace fallout
+
+#endif /* ANIMATION_DEFS_H */
