+++ src/art.cc
@@ -8,7 +8,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "animation.h"
+#include "animation_defs.h"
 #include "art_defs.h"
 #include "content_config.h"
 #include "datafile.h"
