+++ src/combat_ai.cc
@@ -7,6 +7,7 @@
 #include <unordered_set>
 
 #include "actions.h"
+#include "animation_defs.h"
 #include "animation.h"
 #include "art.h"
 #include "color.h"
