+++ src/critter.h
@@ -1,7 +1,7 @@
 #ifndef CRITTER_H
 #define CRITTER_H
 
-#include "animation.h"
+#include "animation_defs.h"
 #include "art.h"
 #include "art_defs.h"
 #include "combat_defs.h"
