+++ src/game_mouse.cc
@@ -11,6 +11,7 @@
 #include <algorithm>
 
 #include "actions.h"
+#include "animation_defs.h"
 #include "animation.h"
 #include "art.h"
 #include "color.h"
