+++ src/item.cc
@@ -6,7 +6,7 @@
 #include <unordered_set>
 #include <vector>
 
-#include "animation.h"
+#include "animation_defs.h"
 #include "art.h"
 #include "art_defs.h"
 #include "automap.h"
