+++ src/mapper/mapper.cc
@@ -8,6 +8,7 @@
 #include <string.h>
 
 #include "actions.h"
+#include "animation_defs.h"
 #include "animation.h"
 #include "art.h"
 #include "automap.h"
