+++ src/object.cc
@@ -6,6 +6,7 @@
 #include <algorithm>
 #include <vector>
 
+#include "animation_defs.h"
 #include "animation.h"
 #include "art.h"
 #include "color.h"
