+++ src/proto_instance.cc
@@ -7,6 +7,7 @@
 
 #include <string>
 
+#include "animation_defs.h"
 #include "animation.h"
 #include "art.h"
 #include "color.h"
