+++ src/sfall_animation.cc
@@ -3,6 +3,7 @@
 #include <algorithm>
 #include <stdint.h>
 
+#include "animation_defs.h"
 #include "animation.h"
 #include "interpreter.h"
 #include "opcode_context.h"
